% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{as_MSnSet}
\alias{as_MSnSet}
\title{Convert and MSnID to an MSnSet}
\usage{
as_MSnSet(x, fcol = NULL)
}
\arguments{
\item{x}{An object of class \code{MSnID}.}

\item{fcol}{Feature variable used to combine PSMs.}
}
\value{
An \code{MSnSet} object.
}
\description{
A function to convert and MSnID to an MSnSet. Used to replacet the MSnID
function \code{as(., "MSnSet")} that drops features. Eventually, this function
will replace the one in MSnID.
}
