% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{dropConstantVariables}
\alias{dropConstantVariables}
\title{Drop constant variable}
\usage{
dropConstantVariables(exp)
}
\arguments{
\item{exp}{An experiment summary, as produced by \code{readSampleExperimentTable}.}
}
\value{
An update experiment table without any constant variables.
}
\description{
Drop constant variable
}
