% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{experimentHierarchy}
\alias{experimentHierarchy}
\title{Print a summary hierarchy.}
\usage{
experimentHierarchy(exp, fcol, orderBy = NULL)
}
\arguments{
\item{exp}{An experiment summary, as produced by \code{readSampleExperimentTable}.}

\item{fcol}{Variable names to be used for the hierarchy/}

\item{orderBy}{Optional variable to order.}
}
\value{
Used for printing only.
}
\description{
Print a summary hierarchy.
}
