% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{experimentSummary}
\alias{experimentSummary}
\title{Print a experiment summary}
\usage{
experimentSummary(exp)
}
\arguments{
\item{exp}{An experiment summary, as produced by \code{readSampleExperimentTable}.}
}
\value{
Used for its side effect.
}
\description{
Print a experiment summary
}
