% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{A polite function}
\usage{
hello2(n = 1)
}
\arguments{
\item{n}{Integer with default value of 1.}
}
\value{
Nothing, used for greating only.
}
\description{
A polite function
}
\examples{
hello2()
}
