% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{readSampleExperimentTable}
\alias{readSampleExperimentTable}
\title{Read the sample experiment table}
\usage{
readSampleExperimentTable(f = "SampleExperimentTable.csv",
  mzml = "mzml", mzid = "msgf")
}
\arguments{
\item{f}{\code{character(1)} defining the name of the
sample experiment table in comma separated values
(csv) format. Default is \code{"SampleExperimentTable.csv"}.}

\item{mzml}{\code{character(1)} defining the directory name
where the mzML files can be found. Default is
\code{"mzml"}.}

\item{mzid}{\code{character(1)} defining the directory name
where the mzid files can be found. Default is
\code{"msgf"}.}
}
\value{
\code{data.frame} with the sample experiment table.
}
\description{
This function read the sample experiment table
and checks that all the mzML and mzid files exist.
}
